Conflicts:
	bank/main.cpp

+++ bank/deleteaccount.cpp
@@ -1,11 +1,14 @@
 #include "deleteaccount.h"
 #include "ui_deleteaccount.h"
+#include <QMessageBox>
 
 DeleteAccount::DeleteAccount(QWidget *parent) :
     QDialog(parent),
     ui(new Ui::DeleteAccount)
 {
     ui->setupUi(this);
+    QRegExp re(QString("^[0-9]+$"));
+    ui->accountLineEdit->setValidator(new QRegExpValidator(re,this));
 }
 
 DeleteAccount::~DeleteAccount()
@@ -15,5 +18,13 @@ DeleteAccount::~DeleteAccount()
 
 void DeleteAccount::on_buttonBox_accepted()
 {
-    b->deleteAccount(ui->accountLineEdit->text());
+    QMessageBox message;
+    if(!ui->accountLineEdit->hasAcceptableInput()){
+        message.setText(QString("输入格式错误"));
+        message.exec();
+        return;
+    }
+    int status = b->deleteAccount(ui->accountLineEdit->text());
+    message.setText(status?QString("不存在此账号"):QString("删除账号成功"));
+    message.exec();
 }
